+++ src/main/scala/org/podval/docbook/gradle/Evaluator.scala
@@ -1,5 +1,0 @@
-package org.podval.docbook.gradle
-
-trait Evaluator {
-  def eval(expression: String): Option[AnyRef]
-}