+++ src/main/scala/org/podval/docbook/gradle/Saxon.scala
@@ -12,20 +12,17 @@ import org.apache.xerces.jaxp.SAXParserFactoryImpl
 import org.xml.sax.{InputSource, XMLReader}
 
 class Saxon(
-  xslParameters: Map[String, String],
   entities: Map[String, String],
+  substitutions: Map[String, String],
   xslDirectory: File,
-  imagesDirectoryName: String,
   dataDirectory: File,
-  evaluator: Evaluator,
   logger: Logger
 ) {
   logger.info(
     s"""Created new Saxon(
-       |  xslParameters = $xslParameters,
        |  entities = $entities,
+       |  substitutions = $substitutions,
        |  xslDirectory = "$xslDirectory",
-       |  imagesDirectoryName = "$imagesDirectoryName",
        |  dataDirectory = "$dataDirectory"
        |)""".stripMargin
   )
@@ -43,7 +40,7 @@ class Saxon(
 
   val xmlReader: XMLReader = new ProcessingInstructionsFilter(
     parent = saxParserFactory.newSAXParser.getXMLReader,
-    evaluator = evaluator,
+    substitutions = entities ++ substitutions,
     logger = logger
   )
   xmlReader.setEntityResolver(entityResolver)
@@ -63,78 +60,36 @@ class Saxon(
   def run(
     inputFile: File,
     stylesheetFile: File,
+    xslParameters: Map[String, String],
     outputFile: File
   ): Unit = run(
     inputSource = new InputSource(inputFile.toURI.toASCIIString),
     stylesheetSource = new StreamSource(stylesheetFile),
-    outputTarget = new StreamResult(outputFile),
-    outputFile = Some(outputFile)
+    xslParameters = xslParameters,
+    outputTarget = new StreamResult(outputFile)
   )
 
   def run(
     inputSource: InputSource,
     stylesheetSource: Source,
-    outputTarget: Result,
-    outputFile: Option[File]
+    xslParameters: Map[String, String],
+    outputTarget: Result
   ): Unit = {
     s"""Saxon.run(
        |  inputSource = ${inputSource.getSystemId},
        |  stylesheetSource = ${stylesheetSource.getSystemId},
-       |  outputFile = $outputFile
+       |  xslParameters = $xslParameters,
+       |  outputTarget = ${outputTarget.getSystemId}
        |)""".stripMargin
 
     logger.info("Configuring Transformer")
     val transformer: Transformer = transformerFactory.newTransformer(stylesheetSource)
 
     logger.info("Setting Transformer parameters")
-    setParameters(transformer, outputFile)
-
-    outputFile.foreach(_.getParentFile.mkdirs)
+    xslParameters.foreach { case (name: String, value: String) => transformer.setParameter(name, value) }
 
     logger.info("Running transform")
     val xmlSource: SAXSource = new SAXSource(xmlReader, inputSource)
     transformer.transform(xmlSource, outputTarget)
   }
-
-  private def setParameters(transformer: Transformer, outputFile: Option[File]): Unit = {
-    def setParameter(name: String, value: String): Unit = {
-      logger.info(s"Transformer parameter $name=$value")
-      transformer.setParameter(name, value)
-    }
-
-    def setOptionalParameter(name: String, value: String): Unit =
-      if (!xslParameters.contains(name)) setParameter(name, value)
-
-    xslParameters.foreach { case (name: String, value: String) => setParameter(name, value) }
-
-    // Relevant only for HTML and EPUB:
-    setOptionalParameter("img.src.path", imagesDirectoryName + "/")
-    setOptionalParameter("html.stylesheet", "css/docBook.css")
-
-    // Relevant only for chunked HTML:
-    outputFile.foreach { outputFile =>
-      setOptionalParameter("root.filename", Util.fileNameWithoutExtension(outputFile))
-      setOptionalParameter("base.dir", outputFile.getParent)
-    }
-  }
-}
-
-object Saxon {
-  def apply(
-    xslParameters: Map[String, String],
-    entities: Map[String, String],
-    xslDirectory: File,
-    imagesDirectoryName: String,
-    dataDirectory: File,
-    evaluator: Evaluator,
-    logger: Logger
-  ): Saxon = new Saxon(
-    xslParameters = xslParameters,
-    entities = entities,
-    xslDirectory = xslDirectory,
-    imagesDirectoryName = imagesDirectoryName,
-    dataDirectory = dataDirectory,
-    evaluator = evaluator,
-    logger = logger
-  )
 }