+++ src/test/scala/org/podval/docbook/gradle/SaxonUtil.scala
@@ -6,14 +6,12 @@ import javax.xml.transform.stream.StreamResult
 import org.xml.sax.InputSource
 
 object SaxonUtil {
-  def toHtml(resourceName: String, entities: Map[String, String]): String = {
+  def toHtml(resourceName: String, substitutions: Map[String, String]): String = {
     val saxon: Saxon = new Saxon(
       xslDirectory = new File("build/docBookXsl/docbook").getAbsoluteFile,
-      xslParameters = Map(),
-      entities = entities,
-      imagesDirectoryName = "images",
+      entities = Map.empty,
+      substitutions = substitutions,
       dataDirectory = new File("/tmp/data"),
-      evaluator = getEvaluator(entities),
       logger = new Logger.TestLogger
     )
 
@@ -23,15 +21,10 @@ object SaxonUtil {
       inputSource = getResourceAsInputSource(resourceName),
       stylesheetSource = saxon.resolve("http://docbook.sourceforge.net/release/xsl-ns/current/html/docbook.xsl"),
       outputTarget = new StreamResult(output),
-      outputFile = None
+      xslParameters = Map()
     )
 
     output.toString
-  }
-
-  private def getEvaluator(properties: Map[String, String]): Evaluator = new Evaluator {
-    override def eval(expression: String): Option[AnyRef] =
-      properties.get(expression)
   }
 
   private def getResourceAsInputSource(name: String): InputSource = {